requirements.txt
transformers>=4.30.0
optimum>=1.13.0
bitsandbytes>=0.41.0
accelerate>=0.21.0
sentencepiece>=0.1.99
tokenizers>=0.13.3
einops>=0.6.1
faiss-cpu>=1.7.4
chromadb>=0.4.13
numpy>=1.24.3
pandas>=2.0.2
onnx>=1.14.0
onnxruntime>=1.15.1
tqdm>=4.65.0
PyYAML>=6.0
rich>=13.4.2
sentence-transformers>=2.2.2
matplotlib>=3.7.1
seaborn>=0.12.2
psutil>=5.9.5